requirements.txt
numpy
torch
pillow
torchvision
matplotlib
#-e .